refactor(pipeline): organized source code. Fixed bugs
- Moved pipeline into root. Kept interfaces in the interfaces subdirectory.
- Fixed bug: aes-192-cbc and aes-256-cbc had invalid IV sizes.
