+++ src/interfaces/IPipeline.ts
